- Unifying LocalShapeFunctionSetProvider and LocalDoFSetProvider to one facility LocalFiniteElementProvider - Introducing DimLocalDoFSet - Deriving LocalShapeFunctionSet from DimLocalDoFSet - FEProvider only creates spaces on request, trys to create SubElement-DoFSets on request, returns SubElement-LSFS only if available and registered (i.e. continuous Spaces only) - Removing hacks for discontinuous Spaces in DoFPositions and InnerDoFPositions - Some cleanup in Implementations of LocalShapeFunctionSets (to be continued ...)

+++ ugbase/lib_disc/dof_manager/surface_dof_distribution.cpp
@@ -880,7 +880,7 @@ get_connections(std::vector<std::vector<size_t> >& vvConnection) const
 //	if this is not the case for non-grouped DoFs, we cannot allow reordering
 	if(!MGDoFDistribution::grouped())
 	{
-		size_t numDoFs = 0;
+		int numDoFs = 0;
 		for(int si = 0; si < num_subsets(); ++si)
 			for(int i = 0; i < NUM_REFERENCE_OBJECTS; ++i)
 			{
